[python][ray] Move incremental target preparation to sink
