+++ paimon-python/pypaimon/write/ray_datasink.py
@@ -263,6 +263,66 @@ class PaimonDatasink(_DatasinkBase):
                 self._pending_commit_messages = []
 
 
+def _prepare_incremental_update(table, target, update_cols):
+    from pypaimon.common.options.core_options import MergeEngine
+    from pypaimon.schema.data_types import PyarrowFieldParser
+    from pypaimon.table.bucket_mode import BucketMode
+
+    if not update_cols:
+        raise ValueError("update_cols must be non-empty.")
+    update_cols = list(dict.fromkeys(update_cols))
+    if not (
+            table.is_primary_key_table
+            and table.bucket_mode() == BucketMode.HASH_FIXED
+            and table.options.merge_engine() == MergeEngine.PARTIAL_UPDATE):
+        raise ValueError(
+            "incremental write_paimon requires a fixed-bucket primary-key "
+            "target with 'merge-engine'='partial-update'.")
+    if table.cross_partition_update:
+        raise ValueError(
+            "incremental write_paimon does not support cross-partition "
+            "updates.")
+    if table.options.sequence_field():
+        raise ValueError(
+            "incremental write_paimon does not support sequence fields.")
+
+    primary_keys = list(table.primary_keys)
+    invalid = [
+        name for name in update_cols
+        if name not in table.field_names or name in primary_keys]
+    if invalid:
+        raise ValueError(
+            "update column {!r} must be a non-key column in target {!r}.".format(
+                invalid[0], target))
+    omitted_non_null = [
+        field.name for field in table.table_schema.fields
+        if (field.name not in primary_keys
+            and field.name not in update_cols
+            and not field.type.nullable)
+    ]
+    if omitted_non_null:
+        raise ValueError(
+            "unprovided partial-update column {!r} must be nullable.".format(
+                omitted_non_null[0]))
+
+    target_schema = PyarrowFieldParser.from_paimon_schema(
+        table.table_schema.fields)
+    required = primary_keys + update_cols
+
+    def to_write_batch(batch):
+        missing = [name for name in required if name not in batch.column_names]
+        if missing:
+            raise ValueError("source is missing columns {}.".format(missing))
+        arrays = [
+            batch.column(field.name).cast(field.type)
+            if field.name in required
+            else pa.nulls(batch.num_rows, type=field.type)
+            for field in target_schema
+        ]
+        return pa.Table.from_arrays(arrays, schema=target_schema)
+    return update_cols, to_write_batch
+
+
 def write_paimon_dataset(
     dataset,
     table,
